requirements.txt
# requirements.txt
discord.py[voice]==2.3.2
yt-dlp==2023.12.30
aiohttp==3.9.1
python-dotenv==1.0.0
PyNaCl==1.3.0
ffmpeg-python==0.2.0